Built the Settings and Channel Pages.
Modified the App.jsx to take route requests from the channel and also nested the settings route.
Created a Header component of the channel's page to use it in the subscription page and in the channel page.
Added the register link to the main navigation bar.
Modified the left navigation bar to include the channel link.
Finalized the user settings with logout possibility, at the front end